Attempting to use the concept count functions again.
