+++ flows/boundary.py
@@ -7,6 +7,7 @@ import json
 import os
 import re
 import tempfile
+from collections import Counter
 from collections.abc import Generator
 from io import BytesIO
 from pathlib import Path
@@ -876,3 +877,210 @@ async def index_by_s3(
                         f"failed to process document batch in indexing task batch #{i}: {str(result)}",
                     )
                     continue
+
+
+def update_s3_with_all_successes(
+    document_object_uri: DocumentObjectUri,
+    cache_bucket: str,
+    concepts_counts_prefix: str,
+) -> None:
+    logger = get_run_logger()
+
+    logger.info("updating S3 with all successes")
+
+    s3 = boto3.client("s3")
+
+    s3_uri = Path(document_object_uri)
+
+    # First, delete the concepts counts object
+    # Get all parts after the prefix (e.g. "Q787/v4/CCLW.executive.1813.2418.json")
+    key_parts = "/".join(s3_uri.parts[3:])  # Skip s3://bucket/labelled_passages/
+
+    concepts_counts_key = f"{concepts_counts_prefix}/{key_parts}"
+
+    _ = s3.delete_object(Bucket=cache_bucket, Key=concepts_counts_key)
+
+    logger.info("updated S3 with deleted concepts counts")
+
+    # Second, delete the labelled passages
+    # Get all parts except for the bucket (e.g. "labelled_passages/Q787/v4/CCLW.executive.1813.2418.json")
+    labelled_passages_key = "/".join(s3_uri.parts[2:])  # Skip s3://bucket/
+
+    _ = s3.delete_object(Bucket=cache_bucket, Key=labelled_passages_key)
+
+    logger.info("updated S3 with deleted labelled passages")
+
+    logger.info("updated S3 with all successes")
+
+    return None
+
+
+def serialise_concepts_counts(concepts_counts: Counter[ConceptModel]) -> str:
+    return json.dumps({str(k): v for k, v in concepts_counts.items()})
+
+
+def save_labelled_passages_by_uri(
+    document_object_uri: DocumentObjectUri,
+    labelled_passages: list[LabelledPassage],
+) -> None:
+    """Save LabelledPassages objects to S3."""
+    object_json = json.dumps(
+        [labelled_passage.model_dump_json() for labelled_passage in labelled_passages]
+    )
+
+    _ = s3_object_write_text(
+        s3_uri=document_object_uri,
+        text=object_json,
+    )
+
+
+def update_s3_with_some_successes(
+    document_object_uri: DocumentObjectUri,
+    concepts_counts_filtered: Counter[ConceptModel],
+    document_labelled_passages: list[LabelledPassage],
+    cache_bucket: str,
+    concepts_counts_prefix: str,
+) -> None:
+    logger = get_run_logger()
+
+    logger.info("updating S3 with partial successes")
+
+    # First, update the concepts counts object
+    serialised_concepts_counts = serialise_concepts_counts(concepts_counts_filtered)
+
+    s3_uri = Path(document_object_uri)
+
+    # Get all parts after the prefix (e.g. "Q787/v4/CCLW.executive.1813.2418.json")
+    key_parts = "/".join(s3_uri.parts[3:])  # Skip s3://bucket/labelled_passages/
+
+    concepts_counts_uri = f"s3://{cache_bucket}/{concepts_counts_prefix}/{key_parts}"
+
+    _ = s3_object_write_text(
+        s3_uri=concepts_counts_uri,
+        text=serialised_concepts_counts,
+    )
+
+    logger.info("updated S3 with updated concepts counts")
+
+    # Second, update the labelled passages
+    concept_ids_to_keep: list[WikibaseID] = [
+        concept_model.wikibase_id for concept_model in concepts_counts_filtered
+    ]
+
+    filtered_labelled_passages: list[LabelledPassage] = []
+
+    for labelled_passage in document_labelled_passages:
+        # It doesn't matter if this list is empty, as it
+        # emulates an empty result from the inference
+        # pipeline.
+        updated_spans: list[Span] = [
+            span
+            for span in labelled_passage.spans
+            if span.concept_id in concept_ids_to_keep
+        ]
+
+        labelled_passage.spans = updated_spans
+
+        filtered_labelled_passages.append(labelled_passage)
+
+    _ = save_labelled_passages_by_uri(
+        document_object_uri=document_object_uri,
+        labelled_passages=filtered_labelled_passages,
+    )
+
+    logger.info("updated S3 with updated labelled passages")
+
+    logger.info("updated S3 with partial successes")
+
+    return None
+
+
+async def update_s3_with_latest_concepts_counts(
+    document_importer: DocumentImporter,
+    concepts_counts: Counter[ConceptModel],
+    cache_bucket: str,
+    concepts_counts_prefix: str,
+    document_labelled_passages: list[LabelledPassage],
+) -> None:
+    logger = get_run_logger()
+
+    # Ideally, we'd remove the concepts count file entirely, but, we may fail above in updating
+    # 1 or more document passages in Vespa, which means that they'd still have the concept present.
+    #
+    # To avoid a mismatch of the family documents' concepts counts, and what's _still_ reflected on
+    # document passages due to failed partial updates, still write an updated concepts counts to
+    # S3.
+    #
+    # However, if we successfully removed all of the concepts from the document passages, then we can
+    # delete it. Then, also update the family document's concepts counts to remove it from there.
+
+    # Remove entries with a value of 0 from the counter
+    concepts_counts_filtered = Counter(
+        {k: v for k, v in concepts_counts.items() if v != 0}
+    )
+
+    # If after filtering out, there's no concepts, that means we
+    # succeeded in all the partial updates to the document
+    # passages.
+    if len(concepts_counts_filtered) == 0:
+        logger.info("successfully updated all concepts")
+        update_s3_with_all_successes(
+            document_object_uri=document_importer[1],
+            cache_bucket=cache_bucket,
+            concepts_counts_prefix=concepts_counts_prefix,
+        )
+    # We didn't succeed with all, so write the concepts counts still
+    else:
+        logger.info("only updated some concepts")
+        update_s3_with_some_successes(
+            document_object_uri=document_importer[1],
+            concepts_counts_filtered=concepts_counts_filtered,
+            document_labelled_passages=document_labelled_passages,
+            cache_bucket=cache_bucket,
+            concepts_counts_prefix=concepts_counts_prefix,
+        )
+
+    return None
+
+
+def calculate_concepts_counts_from_results(
+    results: list[BaseException | None],
+    batch: list[tuple[TextBlockId, list[VespaConcept]]],
+) -> Counter[ConceptModel]:
+    logger = get_run_logger()
+
+    # This can handle multiple concepts, but, in practice at the
+    # moment, this function is operating on a DocumentImporter,
+    # which represents a labelled passages object, which is per
+    # concept.
+    concepts_counts: Counter[ConceptModel] = Counter()
+
+    for i, result in enumerate(results):
+        _text_block_id, concepts = batch[i]
+
+        # Example:
+        #
+        # ..
+        # "labellers": [
+        #   "KeywordClassifier(\"professional services sector\")"
+        # ],
+        # ...
+        concepts_models = [
+            ConceptModel(wikibase_id=WikibaseID(concept.id), model_name=concept.model)
+            for concept in concepts
+        ]
+
+        # Set 0s in the counter for all seen concepts. This ensures
+        # all concepts are represented in the counter even if they're
+        # not updated.
+        for concept_model in concepts_models:
+            if concept_model not in concepts_counts:
+                concepts_counts[concept_model] = 0
+
+        if isinstance(result, Exception):
+            # Since we failed to remove them from the spans, make sure
+            # they're accounted for as remaining.
+            logger.info(f"partial update failed: {str(result)}")
+            concepts_counts.update(concepts_models)
+
+    return concepts_counts
